update response status code

+++ AIBookStreet.API/RequestModel/OrderRequest.cs
@@ -3,7 +3,7 @@
     public class OrderRequest
     {
         public Guid? Id { get; set; }
-        public decimal? TotalAmount { get; set; }
+        public int? TotalAmount { get; set; }
         public string PaymentMethod { get; set; } = null!;
         public DateTime CreatedDate { get; set; }
         public StoreRequest? Store { get; set; }
